rptest: Refactor RedpandaCloud structure/dependencies
  To achieve better code readability and prevent future redpanda.py file
grown to >4k lines, updated dependencies and moved classed to separate
modules. Also, updated cloud cluster detection to
config key GLOBAL_CLOUD_CLUSTER_CONFIG instead of
GLOBAL_CLOUD_OAUTH_URL. RedpandaServiceCloud uses dataclass to serialize
all parameters from globals.json section 'cloud_config'. This will
prevent us from adding numerous GLOBAL_CLOUD_CLUSTER_* vars in the
future.
