+++ tests/rptest/services/rp_config_profiles/redpanda.cloud-tiers-config.yml
